[REFACTOR] implementing separation of concern by separate each test to different file
