+++ authentication/tests/test_authentication.py
@@ -1,0 +1,38 @@
+from .test import *
+
+class AuthenticationTests(TestCase):
+    def setUp(self):
+        self.client = TestClient(router)
+        self.user = User.objects.create_user(username="testuser", email="test@example.com", password="password")
+        self.refresh = RefreshToken.for_user(self.user)
+
+    def test_process_session_existing_user(self):
+        response = self.client.post("/process-session", json={"user": {"email": "test@example.com", "name": "testuser"}})
+        self.assertEqual(response.status_code, 200)
+        self.assertIn("access", response.json())
+
+    def test_process_session_new_user(self):
+        response = self.client.post("/process-session", json={"user": {"email": "new@example.com", "name": "newuser"}})
+        self.assertEqual(response.status_code, 200)
+        self.assertEqual(response.json()["user"]["email"], "new@example.com")
+
+    def test_refresh_token_valid(self):
+        response = self.client.post("/refresh-token", json={"refresh": str(self.refresh)})
+        self.assertEqual(response.status_code, 200)
+        self.assertIn("access", response.json())
+
+    def test_refresh_token_invalid(self):
+        response = self.client.post("/refresh-token", json={"refresh": "invalid_token"})
+        self.assertEqual(response.status_code, 401)
+        self.assertIn("error", response.json())
+
+    def test_validate_token_valid(self):
+        access_token = str(self.refresh.access_token)
+        response = self.client.post("/validate-token", json={"token": access_token})
+        self.assertEqual(response.status_code, 200)
+        self.assertTrue(response.json()["valid"])
+
+    def test_validate_token_invalid(self):
+        response = self.client.post("/validate-token", json={"token": "invalid_token"})
+        self.assertEqual(response.status_code, 200)
+        self.assertFalse(response.json()["valid"])
